requirements.txt
amqp==2.6.1
arrow==0.4.2
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
celery==4.4.0
celery-stubs==0.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-filter==2.4.0
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djangorestframework-stubs==1.4.0
djoser==2.1.0
flower==1.0.0
humanize==3.11.0
ics==0.4
idna==2.10
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
kombu==4.6.11
MarkupSafe==2.0.1
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
TatSu==5.6.1
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==1.3.0
wcwidth==0.2.5
